requirements.txt
flask==1.1.2
Flask-API==2.0
flask-restful==0.3.8
flask_swagger_ui==3.25.0
SQLAlchemy==1.3.20
flask-sqlalchemy==2.4.4
PyMySQL==0.10.1
python-dotenv==0.17.0
werkzeug==1.0.1
requests==2.23.0
pytest==6.2.5
pytest-cov==3.0.0